+++ agate/tableset/having.py
@@ -1,0 +1,42 @@
+#!/usr/bin/env python
+
+from collections import OrderedDict
+
+
+def having(self, aggregations, test):
+    """
+    Create a new :class:`.TableSet` with only those tables that pass a test.
+
+    This works by applying a sequence of :class:`Aggregation` instances to
+    each table. The resulting dictionary of properties is then passed to
+    the :code:`test` function.
+
+    This method does not modify the underlying tables in any way.
+
+    :param aggregations:
+        A list of tuples in the format :code:`(name, aggregation)`, where
+        each :code:`aggregation` is an instance of :class:`.Aggregation`.
+    :param test:
+        A function that takes a dictionary of aggregated properties and returns
+        :code:`True` if it should be included in the new :class:`.TableSet`.
+    :type test:
+        :class:`function`
+    :returns:
+        A new :class:`.TableSet`.
+    """
+    from agate.tableset import TableSet
+    
+    new_tables = []
+    new_keys = []
+
+    for key, table in self.items():
+        props = OrderedDict()
+
+        for name, aggregation in aggregations:
+            props[name] = table.aggregate(aggregation)
+
+        if test(props):
+            new_tables.append(table)
+            new_keys.append(key)
+
+    return TableSet(new_tables, new_keys, self._key_name, self._key_type)